Split the tfvars handling into a separate syntax file
The main syntax no longer handles the .tfvars extension. Handling of
.tfvars is now done by the new syntax file which also assigns a new
source.terraform-vars scope instead of the previous source.terraform.

This change is to allow for differentiating .tfvars and .tf files when
using a terraform-ls language server.

The .tmLanguage syntax files are not changed in the same way as those
don't support including other syntaxes so would have to be duplicated
instead. And since those are long obsolete (in ST at least), I didn't
see much point in doing anything with them.