Revise block matching
Match all blocks with multi-push statements, while still special casing
2-label resource blocks but additionally supporting literal labels.

Contexts are renamed appropriately and variables with identifier "lists"
are made atomic so that they can be used in patterns as atomics.

Meta scopes of blocks are updated so that the entire thing gets
`meta.block` and the body part is specialized to `meta.block.body`.